split large readme into README.md
