+++ trulens_eval/trulens_eval/feedback/provider/cohere.py
@@ -1,5 +1,5 @@
 from trulens_eval.feedback import prompts
-from trulens_eval.feedback.provider import Provider
+from trulens_eval.feedback.provider.base import Provider
 from trulens_eval.feedback.provider.endpoint.base import Endpoint
 from trulens_eval.keys import get_cohere_agent
